Refactored script.js to use a multidimensional "choices" key during callToAction events.
Modified program logic to look for a "choices" key on "CallToAction" nodesin the story.gameText object. The multidimensional array of choices is then used to populate the "act" action available during gameplay. Making a choice in "act" instructs the program to navigate to a new part of the story.gameText object to continue the game. Modifications to the gameText object need to be completed.